tutorials/09-authentication/pyproject.toml
[project]
name = "mcp-authentication-tutorial"
version = "0.1.0"
description = "MCP Server with Authentication Tutorial"
authors = [{name = "MCP Tutorial", email = "tutorial@mcp.dev"}]
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "aiosqlite>=0.19.0",
    "cryptography>=41.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "slowapi>=0.1.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["libs"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.6.0",
]
